Move studio custom logic to separate folder
